# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Architectural Portfolio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: published
          in: query
          schema:
            type: boolean
        - name: category_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /projects/{slug}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
  /projects/{id}/update:
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}/publish:
    patch:
      operationId: toggleProjectPublish
      tags: [projects]
      summary: Toggle project published status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishToggle"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}/delete:
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ProjectImage:
      type: object
      required: [id, projectId, imageUrl, isHero, sortOrder]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        imageUrl:
          type: string
        isHero:
          type: boolean
        sortOrder:
          type: integer
...
    ProjectSummary:
      type: object
      required: [id, title, slug, status, published]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        location:
          type: ["string", "null"]
        client:
          type: ["string", "null"]
        sector:
          type: ["string", "null"]
        status:
          type: string
        published:
          type: boolean
        featured:
          type: boolean
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        heroImage:
          type: ["string", "null"]
        year:
          type: ["string", "null"]
    Project:
      type: object
      required: [id, title, slug, status, published]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        location:
          type: ["string", "null"]
        client:
          type: ["string", "null"]
        sector:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        scope:
          type: ["string", "null"]
        status:
          type: string
        published:
          type: boolean
        featured:
          type: boolean
        longDescription:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        heroImage:
          type: ["string", "null"]
        year:
          type: ["string", "null"]
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProjectImage"
    ProjectInput:
      type: object
      required: [title, slug, status]
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        location:
          type: string
        client:
          type: string
        sector:
          type: string
        size:
          type: string
        scope:
          type: string
        status:
          type: string
        longDescription:
          type: string
        categoryId:
          type: ["integer", "null"]
        year:
          type: string
        published:
          type: boolean
        featured:
          type: boolean
    PublishToggle:
      type: object
      required: [published]
      properties:
        published:
          type: boolean
...