# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Architectural Portfolio API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectImage } from './projectImage';

export interface Project {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  client?: string | null;
  /** @nullable */
  sector?: string | null;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  scope?: string | null;
  status: string;
  published: boolean;
  featured?: boolean;
  /** @nullable */
  longDescription?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  heroImage?: string | null;
  /** @nullable */
  year?: string | null;
  images?: ProjectImage[];
}